Changes to the api and datastructure:
- Move from clientside filtering to server side filtering
- Filter privatePartitionData and scratchStorageData based on group membership
  instead of username
- Add env variable to keep track of the app basename
-

Serverside filtering means we will be able to send smaller files over the
network. It also means we can abstract away the frontend more and just modify
the backend code when changing which data gets displayed on the frontend.

It was decided to filter based on group membership because users are given
access to a lab based on group membership. This means we don't have to update
the username information as often; if a user is added to the group he/she is
good to go.

The env variable will make it easier to change between routes. It will need to
be changed for production to the production subroute.